# Improve BitArray's Get, Set, SetAll and Not methods performance