@Wraith2 2019-02-16
AggressiveInlining is great for perf but do we want it inlining everywhere that it's used when it might increase the calling method hugely if it's used a lot?