@omariom 2019-02-16
> AggressiveInlining is great for perf but do we want it inlining everywhere that it's used when it might increase the calling method hugely if it's used a lot?

The generated code is  fairly small and improvements, imo, justify inlining.
I haven't shown it in the results but `bitArray.Set(index, true)` is 0.5252. It is about 3.44 times faster than the original version and 2.83 times than non-inlining one. Inlining allowed JIT to remove the check and  `false` branch.