@stephentoub 2019-02-20
> The generated code is fairly small and improvements, imo, justify inlining.

Maybe, but it's still a sizeable increase.  For example, this:
```C#
var ba = new BitArray(10);
for (int i = 0; i < 10; i++) ba[i] = true;
for (int i = 0; i < 10; i++) ba[i] = !ba[i];
```
goes from 111 bytes of asm before the change to 248 after, so an ~2.2x increase.

@jkotas, do you have an opinion on the AggressiveInlining usage here?