@jkotas 2019-02-21
I think the AggressiveInlining on this one is ok.