@jkotas 2019-02-21
E.g. We have AggressiveInlining on `List<T>.Add` that is more painful trade-off.